feat: add code for extracting 1d lines from a 2d fourier transform pt.2
