+++ src/torch_fourier_slice/grids/__init__.py
@@ -1,2 +1,2 @@
-from .central_slice_fftfreq_grid import central_slice_fftfreq_grid
 from .central_line_fftfreq_grid import central_line_fftfreq_grid
+from .central_slice_fftfreq_grid import central_slice_fftfreq_grid
