+++ src/torch_fourier_slice/grids/central_line_fftfreq_grid.py
@@ -1,7 +1,7 @@
 import einops
 import torch
 
-from ..dft_utils import rfft_shape, fftshift_1d
+from ..dft_utils import fftshift_1d, rfft_shape
 
 
 def central_line_fftfreq_grid(
@@ -23,11 +23,11 @@ def central_line_fftfreq_grid(
         zeros = torch.zeros(size=rfft_shape((w,)), dtype=grid.dtype, device=device)
     else:
         zeros = torch.zeros(size=(w,), dtype=grid.dtype, device=device)
-    central_slice_grid, _ = einops.pack([zeros, grid], pattern='w *')  # (w, 2)
+    central_slice_grid, _ = einops.pack([zeros, grid], pattern="w *")  # (w, 2)
 
     # fftshift if requested
     if fftshift is True:
-        central_slice_grid = einops.rearrange(central_slice_grid, 'w freq -> freq w')
+        central_slice_grid = einops.rearrange(central_slice_grid, "w freq -> freq w")
         central_slice_grid = fftshift_1d(central_slice_grid, rfft=rfft)
-        central_slice_grid = einops.rearrange(central_slice_grid, 'freq w -> w freq')
+        central_slice_grid = einops.rearrange(central_slice_grid, "freq w -> w freq")
     return central_slice_grid
